+++ vault/deprecated.py
@@ -1,0 +1,1 @@
+"""This file generates warnings about methods and classes that will be removed in future versions"""